requirements.txt
customtkinter
Pillow
pyheif
tqdm
